Time: 0 ms (100%), Space: 8.5 MB (54.27%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
-    int reverse(long long x) {
-        long long rev =0;
+    int reverse(int x) {
+        int rev =0;
         while(x!=0){
             int digit = x % 10;
             x /= 10;
